second commit

+++ utils/db_api/db_commands.py
@@ -3,9 +3,27 @@ import sqlite3
 def add_user_to_db(user_id, username, fullname):
     conn = sqlite3.connect('bot.db')
     cursor = conn.cursor()
+
+    # Check if user_id exists in the database and retrieve the is_active status
     cursor.execute('''
-    INSERT OR IGNORE INTO users (user_id, username, fullname) VALUES (?, ?, ?)
-    ''', (user_id, username, fullname))
+    SELECT is_active FROM users WHERE user_id = ?
+    ''', (user_id,))
+    result = cursor.fetchone()
+
+    if result is not None:
+        # If user_id exists, check the is_active status
+        if result[0] == 0:  # Assuming 0 is false
+            cursor.execute('''
+            UPDATE users
+            SET is_active = 1
+            WHERE user_id = ?
+            ''', (user_id,))
+    else:
+        # If user_id does not exist, insert the new user
+        cursor.execute('''
+        INSERT OR IGNORE INTO users (user_id, username, fullname, is_active) VALUES (?, ?, ?, 1)
+        ''', (user_id, username, fullname))
+
     conn.commit()
     conn.close()
 
@@ -77,11 +95,97 @@ def count_user():
     else:
         return None
     
+def get_all_user():
+    conn = sqlite3.connect('bot.db')
+    cursor = conn.cursor()
+    cursor.execute('''
+    SELECT user_id FROM users;
+    ''')  
+    result = cursor.fetchall()
+    conn.close()
+
+    return result
+
+def set_inactive(user_id):
+    if not isinstance(user_id, int):
+        raise ValueError("user_id must be an integer")
+
+    conn = sqlite3.connect('bot.db')
+    cursor = conn.cursor()
+    try:
+        cursor.execute('''
+        UPDATE users
+        SET is_active = 0
+        WHERE user_id = ?
+        ''', (user_id,))
+        conn.commit()
+    except sqlite3.Error as e:
+        print(f"An error occurred: {e}")
+    finally:
+        conn.close()
+
+def active_users():
+    conn = sqlite3.connect('bot.db')
+    cursor = conn.cursor()
+    cursor.execute('''
+    SELECT COUNT(*) FROM users WHERE is_active = 1;
+    ''')  
+    result = cursor.fetchone()
+    conn.close()
+    if result:
+        return result[0]
+    else:
+        return None
+
+def inactive_users():
+    conn = sqlite3.connect('bot.db')
+    cursor = conn.cursor()
+    cursor.execute('''
+    SELECT COUNT(*) FROM users WHERE is_active = 0;
+    ''')  
+    result = cursor.fetchone()
+    conn.close()
+    if result:
+        return result[0]
+    else:
+        return None
+
+ 
+def count_videos():
+    conn = sqlite3.connect('bot.db')
+    cursor = conn.cursor()
+    cursor.execute('''
+    SELECT COUNT(*) FROM videos;
+    ''')  
+    result = cursor.fetchone()
+    conn.close()
+    if result:
+        return result[0]
+    else:
+        return None   
+
+def delete_video(id):
+    conn = sqlite3.connect('bot.db')
+    cursor = conn.cursor()
+    cursor.execute('''
+    DELETE FROM videos WHERE id = ?
+    ''', (id,))
+    conn.commit()
+    conn.close()
+
 if __name__ == "__main__":
     conn = sqlite3.connect('bot.db')
     cursor = conn.cursor()
-    cursor.execute('''DROP TABLE videos ''')
-    cursor.execute('''DROP TABLE users ''')
+    ask = input("Do you want to delete all VIDEOS(yes/no): ")
+    if ask == "yes":
+        cursor.execute('''DROP TABLE videos ''')
+        print("All vidoes deleted successfully")
+    
+    ask = input("Do you want to delete all USERS(yes/no): ")
+    if ask == "yes":
+        cursor.execute('''DROP TABLE users ''')
+        print("All users deleted successfully")
+
     conn.commit()
     conn.close()
